Fetch plant details dynamically
